Fix issue where dental_practices not loading
When the etl is run at the theme level, the dental_practices were not
loading.  When run as health.dental_practices it was fine.  The issue
was when we zip up the entrypoints, and the service manifests they were
not matching because the ordering was different in the two locations.
Moving dentists in the manifest to after pharmacists fixed this.